pyproject.toml
[project]
name = "ext"
version = "0.1.0"
description = "抖音字幕提取工具"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.4.0",
    "redis>=5.0.0",
    "pymysql>=1.1.0",
    "imagehash>=4.3.1",
    "Pillow>=10.0.0",
    "faster-whisper>=1.0.0",
    "opencv-python>=4.8.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "loguru>=0.7.0",
    "schedule>=1.2.0",
    "ffmpeg-python>=0.2.0",
    "aiomysql>=0.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
